refactor(test): split the record_review_pass CLI suite along its own seams
Trimming the file back to 495 lines was not enough. The bloat gate blocks on
`delta == "crossing" and not in_baseline` — a property of the file being over
its ceiling at all, not of it having grown in this change. A 496-line file
against a 300-line limit with no grandfathering entry blocks whoever touches
it, and "it was already like that" is exactly the deferral the Iron Law names.

Split along the section headers the file already carried, so the seams are the
ones its author drew rather than ones invented for the line count:

  _review_cli_harness.py                96   fixture builder, subprocess
                                             wrapper, reviewer payloads
  test_record_review_pass_cli.py       192   AC8 / AC7 / AC3 / AC2
  ..._cli_regressions.py               171   the code-review round's findings
  ..._cli_doubt.py                     160   the Stage-3 doubt pass's findings
  ..._cli_floor.py                     113   close-missing x the floor

29 tests before, 29 after; the full shared suite is unchanged at 5895.

Two deliberate choices. The harness is a plain module, not a conftest.py:
shared/tests/conftest.py loads for the whole suite, and a `project` fixture
visible to ~5900 unrelated tests is a collision waiting to happen. And it
exports make_project() as a function rather than a fixture, because importing
a fixture makes ruff read every test taking `project` as an F811 redefinition
of the imported name — 28 of them — which is noisier to silence than a
three-line wrapper per file costs.

No behaviour change; every moved case asserts what it did before.

Spec Impact: NONE — test-file organisation only.

Run-ID: iterate-2026-07-27-split-review-cli-tests
